% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/changeBase.R
\name{changeBase}
\alias{changeBase}
\title{Convert number from one base to another.}
\usage{
changeBase(x, from = NULL, to = 2)
}
\arguments{
\item{x:}{Object of class \code{DecomposedNumber}.}
}
\description{
Convert number from one base to another. Base values are picked between 2 and 36.
}
\examples{
changeBase(-265, from=10, to=16)
changeBase("e2792a", from=16, to=10)
}

