% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/composeNumber.R
\name{composeNumber}
\alias{composeNumber}
\title{Compose an objects of class DecomposedNumber.}
\usage{
composeNumber(x)
}
\arguments{
\item{x}{A number or a character string to be passed to function \code{decomposeNumber} function.}
}
\description{
Convert a \code{DecomposedNumber} object into a object of class \code{numeric} or \code{character}.
}
\examples{
dnum <- DecomposedNumber("-776846,3")
composeNumber(dnum)
}

