% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DecomposedNumber.R
\name{DecomposedNumber}
\alias{DecomposedNumber}
\title{Create objects of class DecomposedNumber.}
\usage{
DecomposedNumber(x, base = NULL)
}
\arguments{
\item{x}{A number or a character string to be passed to function \code{decomposeNumber} function.}
}
\description{
Create objects of class DecomposedNumber from a number or an adequate character string.
}
\examples{
dnum <- DecomposedNumber(-536836.3)
print(dnum)
}

