% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decomposeNumber.R
\name{decomposeNumber}
\alias{decomposeNumber}
\title{Decompose a number.}
\usage{
decomposeNumber(x, signout = FALSE)
}
\arguments{
\item{x}{A number or a character string that contains digits as alphanumeric characters and optionnaly a sign mark and a decimal mark.}

\item{signout}{logical. If TRUE the sign of the number is returned.}
}
\description{
Break apart a multidigit number and return a matrix of its components.
}
\details{
If a sign mark is detected, it will be removed from the matrix returned but retuned if \code{signout} is TRUE.
}
\examples{
decomposeNumber(983)
decomposeNumber("34,53")
decomposeNumber("092092EKXEKE")
decomposeNumber(-536836.3, signout=TRUE)
}

