% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/digitused.R
\name{digitused}
\alias{digitused}
\title{Convert the values of digits used.}
\usage{
digitused(x, inv = FALSE)
}
\arguments{
\item{x}{Either a numeric of a string.}

\item{inv}{logical. If TRUE, the reverse transformed is computed.}
}
\description{
Convert the values of digits used.
}
\examples{
digitused("5", inv=TRUE)
}

