% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.DecomposedNumer.R
\name{print.DecomposedNumber}
\alias{print.DecomposedNumber}
\title{Print methods for DecomposedNumber objects}
\usage{
\method{print}{DecomposedNumber}(x)
}
\arguments{
\item{x:}{Object of class \code{DecomposedNumber}.}
}
\description{
Printing objects of class \code{DecomposedNumber} by simple \code{print} methods.
}
\examples{
dnum <- DecomposedNumber("-536836,3")
print(dnum)
}

