% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.DecomposedNumber.R
\name{summary.DecomposedNumber}
\alias{summary.DecomposedNumber}
\title{Summary methods for DecomposedNumber objects.}
\usage{
\method{summary}{DecomposedNumber}(x)
}
\arguments{
\item{x:}{Object of class \code{DecomposedNumber}.}
}
\description{
Summarize the property of objects of class \code{DecomposedNumber} by simple
\code{print} methods.
}
\examples{
dnum <- DecomposedNumber("-536836,3")
summary(dnum)
}

